Organize the ipfs magic under ipfsInterface
