+++ src/ipfsInterface/adapters/dispute/v1-dispute-adapter.js
@@ -1,0 +1,3 @@
+import AdapterBase from '../adapter-base'
+
+export default class DisputeAdapterV1 extends AdapterBase {}
