+++ src/ipfsInterface/adapters/listing-withdraw/v1-listing-withdraw-adapter.js
@@ -1,0 +1,3 @@
+import AdapterBase from '../adapter-base'
+
+export default class ListingWithdrawAdapterV1 extends AdapterBase {}
