+++ src/ipfsInterface/adapters/offer-accept/v1-offer-accept-adapter.js
@@ -1,0 +1,3 @@
+import AdapterBase from '../adapter-base'
+
+export default class OfferAcceptAdapterV1 extends AdapterBase {}
