+++ src/ipfsInterface/adapters/offer-withdraw/v1-offer-withdraw-adapter.js
@@ -1,0 +1,3 @@
+import AdapterBase from '../adapter-base'
+
+export default class OfferWithdrawAdapterV1 extends AdapterBase {}
