+++ src/ipfsInterface/adapters/profile/v1-profile-adapter.js
@@ -1,0 +1,3 @@
+import AdapterBase from '../adapter-base'
+
+export default class ProfileAdapterV1 extends AdapterBase {}
