+++ src/ipfsInterface/adapters/resolution/v1-resolution-adapter.js
@@ -1,0 +1,3 @@
+import AdapterBase from '../adapter-base'
+
+export default class ResolutionAdapterV1 extends AdapterBase {}
