+++ src/ipfsInterface/adapters/review/v1-review-adapter.js
@@ -1,0 +1,3 @@
+import AdapterBase from '../adapter-base'
+
+export default class ReviewAdapterV1 extends AdapterBase {}
