+++ src/ipfsInterface/schemas/offer-accept.json
