+++ src/ipfsInterface/schemas/offer.json
